homework-7 Добавил ConsoleService

+++ homework7/src/main/java/ru/otus/valeev/controller/LibraryController.java
@@ -1,6 +1,7 @@
 package ru.otus.valeev.controller;
 
 import lombok.AllArgsConstructor;
+import org.apache.commons.lang3.StringUtils;
 import org.springframework.shell.standard.ShellComponent;
 import org.springframework.shell.standard.ShellMethod;
 import org.springframework.shell.standard.ShellOption;
@@ -10,6 +11,7 @@ import ru.otus.valeev.domain.Book;
 import ru.otus.valeev.domain.Genre;
 import ru.otus.valeev.service.AuthorService;
 import ru.otus.valeev.service.BookService;
+import ru.otus.valeev.service.ConsoleService;
 import ru.otus.valeev.service.GenreService;
 
 import java.util.List;
@@ -21,6 +23,7 @@ public class LibraryController {
     private final AuthorService authorService;
     private final BookService bookService;
     private final GenreService genreService;
+    private final ConsoleService consoleService;
 
     @ShellMethod(value = "Get all books", key = {"aa", "allAuthors"})
     public List<Author> getAuthors() {
@@ -38,9 +41,13 @@ public class LibraryController {
     }
 
     @ShellMethod(value = "Create/Save book", key = {"cb", "sb", "createBook", "saveBook"})
-    public Book saveBook(@ShellOption String bookName,
-                         @ShellOption String authorName,
-                         @ShellOption String genreName) {
+    public Book saveBook(@ShellOption(defaultValue = "") String bookName,
+                         @ShellOption(defaultValue = "") String authorName,
+                         @ShellOption(defaultValue = "") String genreName) {
+        if (StringUtils.isBlank(bookName) || StringUtils.isBlank(authorName) || StringUtils.isBlank(genreName)) {
+            consoleService.sendMessage("Ошибка во входных данных при создании книги");
+            return null;
+        }
         return bookService.saveBook(bookName, authorName, genreName);
     }
 
@@ -48,6 +55,10 @@ public class LibraryController {
     public String updateBook(@ShellOption(defaultValue = "") String bookName,
                              @ShellOption(defaultValue = "") String authorName,
                              @ShellOption(defaultValue = "") String genreName) {
+        if (StringUtils.isBlank(bookName) || StringUtils.isBlank(authorName) || StringUtils.isBlank(genreName)) {
+            consoleService.sendMessage("Ошибка во входных данных для изменении книги");
+            return null;
+        }
         Book book = bookService.updateBook(bookName, authorName, genreName);
         if (book != null) {
             return book.toString();
@@ -57,11 +68,18 @@ public class LibraryController {
     }
 
     @ShellMethod(value = "Delete book", key = {"db", "deleteBook"})
-    public String deleteBook(@ShellOption(defaultValue = "") String bookName) {
-        if (bookService.deleteBook(bookName)) {
-            return "Успешное удаление книги с названием: " + bookName;
-        } else {
-            return "Ошибка при удаление книги с названием: " + bookName;
+    public void deleteBook(@ShellOption(defaultValue = "") String bookName) {
+        String message;
+        if (StringUtils.isBlank(bookName)) {
+            message = "Ошибка во входных данных при удалении книги";
+            consoleService.sendMessage(message);
+            return;
         }
+        if (bookService.deleteBook(bookName)) {
+            message = "Успешное удаление книги с названием: " + bookName;
+        } else {
+            message = "Ошибка при удаление книги с названием: " + bookName;
+        }
+        consoleService.sendMessage(message);
     }
 }
